Fix compilation issue

+++ opensync-plugin-0.4x/src/vevent.cc
@@ -813,7 +813,7 @@ bool VEventConverter::CommitRecordData(BarryEnvironment *env, unsigned int dbId,
 		oss << "unable to parse change data for new RecordId: "
 		    << newRecordId << " data: " << data;
 		errmsg = oss.str();
-		trace.logf(errmsg.c_str());
+		trace.log(errmsg.c_str());
 		return false;
 	}
