+++ opensync-plugin-0.4x/src/vjournal.cc
@@ -305,7 +305,7 @@ bool VJournalConverter::CommitRecordData(BarryEnvironment *env, unsigned int dbI
 		oss << "unable to parse change data for new RecordId: "
 		    << newRecordId << " data: " << data;
 		errmsg = oss.str();
-		trace.logf(errmsg.c_str());
+		trace.log(errmsg.c_str());
 		return false;
 	}
